requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.3
python-dotenv==1.0.1
psycopg2-binary==2.9.10
requests==2.32.3
itsdangerous==2.2.0
cryptography>=43.0.0
reportlab==4.2.5
python-docx==1.1.2
python-magic-bin==0.4.14
bleach==6.1.0
cachetools==5.5.0

